Remove priority
This doesn't completely remove it, but it only keeps those pieces
necessary for interoperability with existing implementations: the rules
for sending PRIORITY frames in different stream states, the format of
the HEADERS and PRIORITY frames, and the error handling.

It replaces the exposition about how priority works with a deprecation
notice and the little advice that we currently understand to be
accurate.

Closes #773.